fix(hand tracking): fix issue when hand lose tracking and stays visible
Fixes #273

Me a @saitonakamura checked how the events were sent and found some issues with code not cleaning up, not disposing used resources and not hiding hands when they are not tracking. We decided to add OculusHandModel and XRHandMeshModel to this library so the fix is ready for usage as soon as possible.